% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RcppExports.R
\name{ainv}
\alias{ainv}
\alias{apinv}
\title{Armadillo's Inverse Functions}
\usage{
ainv(x)

apinv(x)
}
\arguments{
\item{x}{a numeric matrix, must be square for \code{ainv}.}
}
\value{
The matrix-inverse or pseudo-inverse.
}
\description{
Matrix inverse and pseudo-inverse by the Armadillo C++ library.
}
\examples{
ainv(crossprod(diff(EuStockMarkets)))

}
