% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.dfm}
\alias{as.data.frame.dfm}
\title{Extract Factor Estimates in a Data Frame}
\usage{
\method{as.data.frame}{dfm}(
  x,
  ...,
  method = "all",
  pivot = c("long", "wide.factor", "wide.method", "wide", "t.wide"),
  time = seq_row(x$F_pca),
  stringsAsFactors = TRUE
)
}
\arguments{
\item{x}{an object class 'dfm'.}

\item{\dots}{not used.}

\item{method}{character. The factor estimates to use: any of \code{"qml"}, \code{"twostep"}, \code{"pca"} (multiple can be supplied) or \code{"all"} for all estimates.}

\item{pivot}{character. The orientation of the frame: \code{"long"}, \code{"wide.factor"} or \code{"wide.method"} or \code{"wide"}.}

\item{time}{a vector identifying the time dimension, or \code{NULL} to omit a time variable.}

\item{stringsAsFactors}{make factors from method and factor identifiers. Same as option to \code{\link{as.data.frame.table}}.}
}
\description{
Extract Factor Estimates in a Data Frame
}
