% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFM.R
\name{DFM}
\alias{DFM}
\title{Estimate a Dynamic Factor Model}
\usage{
DFM(X, r, p, max.iter = 100, tol = 1e-04, miss.threshold = 0.8)
}
\arguments{
\item[r]{number of factors}

\item[p]{number of lags in factor VAR}

\item[tol]{EM convergence tolerance}

\item[miss.threshold]{proportions of variables missing for a case to be removed}

\item[max.inter]{maximum number of EM iterations}
}
\description{
Estimate a Dynamic Factor Model with arbitrary patterns of missing data
on stationary data of a single frequency, with time-invariant system matrices
and idiosynchratic measurement (observation) errors.
}
\details{
The State-Space Model estimated is:

Xt = C Ft + et ~ N(0, R)\cr
Ft = A Ft-1 + ut ~ N(0, Q)

Where
\itemize{
\item{n}{Number of variables}
\item{T}{Number of observations}
\item{F}{Factor vector (stacked, starting with all factors at time t, followed by t-1 etc., rp x 1)}
\item{C}{Observation matrix (n x rp, only the first n x r terms are non-zero (no relationship with lagged factors))}
\item{A}{State transition matrix (rp x rp, lower parts are known (identity matrix))}
\item{Q}{State covariance matrix (rp x rp, top r x r part is contemporaneous, rest is 0 (we assume all relationship between lagged factors is captured in A))}
\item{R}{Observation covariance matrix (n x n, diagonal by assumption)}
}
}
