% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfms.R
\docType{package}
\name{dfms-package}
\alias{dfms-package}
\alias{dfms}
\title{Dynamic Factor Models}
\description{
\emph{dfms} provides efficient estimation of Dynamic Factor Models via the EM Algorithm.
}
\details{
Estimation can be done in 3 different ways following:
\itemize{
\item Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large approximate dynamic factor models based on Kalman filtering. \emph{Journal of Econometrics, 164}(1), 188-205. \url{doi:10.1016/j.jeconom.2011.02.012}
\item Doz, C., Giannone, D., & Reichlin, L. (2012). A quasi-maximum likelihood approach for large, approximate dynamic factor models. \emph{Review of Economics and Statistics, 94}(4), 1014-1024. \url{doi:10.1162/REST_a_00225}
\item Banbura, M., & Modugno, M. (2014). Maximum likelihood estimation of factor models on datasets with arbitrary pattern of missing data. \emph{Journal of Applied Econometrics, 29}(1), 133-160. \url{doi:10.1002/jae.2306}
}

The default is \code{em.method = "auto"}, which chooses \code{"BM"} following Banbura & Modugno (2014) with missing data or mixed frequency, and \code{"DGR"} following Doz, Giannone & Reichlin (2012) otherwise. Using \code{em.method = "none"} generates Two-Step estimates following Doz, Giannone & Reichlin (2011). This is extremely efficient on bigger datasets. PCA and Two-Step estimates are also reported in EM-estimation. All methods support missing data, but \code{em.method = "DGR"} does not model them in EM iterations.
}
\section{Package Contents}{


\strong{Functions to Specify/Estimate Model and Key Methods}

\code{\link[=ICr]{ICr()}} --- Information Criteria\cr
\itemize{
\item \code{\link[=plot.ICr]{plot(<ICr>)}}\cr
\item \code{\link[=screeplot.ICr]{screeplot(<ICr>)}}\cr
}

\code{\link[=DFM]{DFM()}} --- Estimate the Model\cr
\itemize{
\item \code{\link[=summary.dfm]{summary(<dfm>)}}\cr
\item \code{\link[=plot.dfm]{plot(<dfm>)}}\cr
\item \code{\link[=as.data.frame.dfm]{as.data.frame(<dfm>)}}\cr
\item \code{\link[=residuals.dfm]{residuals(<dfm>)}}\cr
\item \code{\link[=fitted.dfm]{fitted(<dfm>)}}
}

\code{\link[=predict.dfm]{predict(<dfm>)}} --- Generate Forecasts\cr
\itemize{
\item \code{\link[=plot.dfm_forecast]{plot(<dfm_forecast>)}}\cr
\item \code{\link[=as.data.frame.dfm_forecast]{as.data.frame(<dfm_forecast>)}}\cr
}

\strong{Auxiliary Functions}

\code{\link[=.VAR]{.VAR()}} --- Estimate Vector Autoregression\cr
\code{\link[=SKF]{SKF()}} --- Stationary Kalman Filter\cr
\code{\link[=FIS]{FIS()}} --- Fixed Interval Smoother\cr
\code{\link[=SKFS]{SKFS()}} --- Stationary Kalman Filter + Smoother\cr
\code{\link[=tsnarmimp]{tsnarmimp()}} --- Remove and Impute Missing Values in a Multivariate Time Series\cr
\code{\link[=ainv]{ainv()}} --- Rcpp Armadillo's Inverse Function\cr
\code{\link[=apinv]{apinv()}} --- Rcpp Armadillo's Pseudo-Inverse Function\cr

\strong{Data}

\code{\link{BM14_M}} --- Monthly Series by Banbura and Modugno (2014)\cr
\code{\link{BM14_Q}} --- Quarterly Series by Banbura and Modugno (2014)\cr
\code{\link{BM14_Models}} --- Series Metadata + Small/Medium/Large Model Specifications\cr
}

