% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.dfm}
\alias{fitted.dfm}
\title{Fitted Values from DFM}
\usage{
\method{fitted}{dfm}(object, method = "qml", orig.format = FALSE, standardized = FALSE, ...)
}
\arguments{
\item{object}{object of class 'dfm'.}

\item{method}{character. \code{"qml"}, \code{"twostep"} or \code{"pca"}.}

\item{orig.format}{logical. \code{TRUE} returns residuals in a data format similar to \code{X}.}

\item{standardized}{logical. \code{FALSE} will put residuals on the original data scale.}
}
\description{
Fitted Values from DFM
}
