% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RcppExports.R
\name{fKFS}
\alias{fKFS}
\title{Fast Kalman Filter and Smoother}
\usage{
fKFS(X, A, C, Q, R, F_0, P_0, loglik = 0L)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{A}{Transition matrix (rp x rp)}

\item{C}{Observation matrix (n x rp)}

\item{Q}{State covariance (rp x rp)}

\item{R}{Observation covariance (n x n)}

\item{F_0}{Initial state vector (rp x 1)}

\item{P_0}{Initial state covariance (rp x rp)}

\item{loglik}{integer. 0 does not compute the likelihood, 1 computes a standard Kalman Filter likelihood, 2 computes the likelihood for Banbura and Modungo (2014).}
}
\value{
All results from \code{\link{fKF}} and \code{\link{fKS}}, and additionally
a rp x rp x T matrix \code{PPm_smooth}, which is equal to the estimate of Cov(F_smooth_t, F_smooth_t-1|T) and needed for EM iterations.
}
\description{
Fast Kalman Filter and Smoother
}
