% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ICr}
\alias{ICr}
\alias{print.ICr}
\alias{plot.ICr}
\alias{screeplot.ICr}
\title{Information Criteria to Determine the Number of Factors (r)}
\usage{
ICr(X, max.r = min(20, ncol(X)))

\method{print}{ICr}(x, ...)

\method{plot}{ICr}(x, ...)

\method{screeplot}{ICr}(x, type = c("pve", "cum.pve"), show.grid = TRUE, max.r = 30, ...)
}
\arguments{
\item{X}{a \code{T x n} data matrix or frame.}

\item{max.r}{integer. The maximum number of factors for which ICs should be computed (or eigenvalues to be displayed in the screeplot).}

\item{type}{character. Either \code{"pve"} (percent variance explained), \code{"cum.pve"} or both.}

\item{show.grid}{logical. \code{TRUE} shows gridlines in each plot.}
}
\value{
A list of 3 elements:
\item{F_pca}{\code{T x n} matrix of principle component factor estimates.}
\item{eigenvalues}{the eigenvalues of the covariance matrix of \code{X}.}
\item{IC}{\code{r.max x 3} 'table' containing the 3 information criteria of Bai & Ng (2002), computed for all values of \code{r} from \code{1:r.max}.}
\item{r.star}{vector of length 3 containing the number of factors (\code{r}) minimizing each information criterion.}
}
\description{
Minimizes 3 information criterio proposed by Bai & Ng (2002) to determine the optimal number of factors r* to be used in an approximate factor model.
A screeplot can also be computed to eyeball the number of factors in the spirit of Onatski (2010).
}
\note{
To determine the number of lags (\code{p}) in the factor transition equation, use the function \code{vars::VARselect} with r* principle components (also returned by \code{ICr}).
}
\references{
Bai, J., Ng, S. (2002). Determining the Number of Factors in Approximate Factor Models. \emph{Econometrica, 70}(1), 191-221. <doi:10.1111/1468-0262.00273>
Onatski, A. (2010). Determining the number of factors from empirical distribution of eigenvalues. \emph{The Review of Economics and Statistics, 92}(4), 1004-1016.
De Valk, S., de Mattos, D., & Ferreira, P. (2019). Nowcasting: An R package for predicting economic variables using dynamic factor models. \emph{The R Journal, 11}(1), 230-244.
}
