% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/my_RcppExports.R
\name{KalmanFilter}
\alias{KalmanFilter}
\title{Implementation of a Kalman filter}
\usage{
KalmanFilter(X, H, Q, R, F, F0, P0)

KalmanFilter(X, H, Q, R, F, F0, P0)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{H}{Observation matrix}

\item{Q}{State covariance}

\item{R}{Observation covariance}

\item{F}{Transition matrix}

\item{F0}{Initial state vector}

\item{P0}{Initial state covariance}

\item{C}{Observation matrix}

\item{A}{Transition matrix}
}
\description{
Implementation of a Kalman filter

Implementation of a Kalman filter
}
