% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/my_RcppExports.R
\name{KalmanSmoother}
\alias{KalmanSmoother}
\title{Runs a Kalman smoother}
\usage{
KalmanSmoother(F, H, R, FfT, FpT, PfT_v, PpT_v)

KalmanSmoother(F, H, R, FfT, FpT, PfT_v, PpT_v)
}
\arguments{
\item{F}{transition matrix}

\item{H}{observation matrix}

\item{R}{Observation covariance}

\item{FfT}{State estimates}

\item{PfT_v}{Variance estimates}

\item{PpT_v}{Predicted variance estimates}

\item{A}{transition matrix}

\item{C}{observation matrix}

\item{FT}{State estimates}

\item{PTm}{State predicted estimates}

\item{FpTm}{State predicted estimates}
}
\value{
List of smoothed estimates

List of smoothed estimates
}
\description{
Runs a Kalman smoother

Runs a Kalman smoother
}
