% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.dfm}
\alias{plot.dfm}
\title{Plot DFM}
\usage{
\method{plot}{dfm}(
  x,
  method = switch(x$em.method, none = "twostep", "qml"),
  type = c("joint", "individual", "residual"),
  ...
)
}
\arguments{
\item{x}{an object class 'dfm'.}

\item{method}{character. The factor estimates to use: one of \code{"qml"}, \code{"twostep"}, \code{"pca"} or \code{"all"} to plot all estimates.}

\item{type}{character. The type of plot: \code{"joint"}, \code{"individual"} or \code{"residual"}.}

\item{\dots}{further arguments to \code{\link{plot}}, \code{\link{ts.plot}}, or \code{\link{boxplot}}, depending on the \code{type} of plot.}
}
\description{
Plot DFM
}
