% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.dfm}
\alias{predict.dfm}
\title{DFM Forecasts}
\usage{
\method{predict}{dfm}(object, h = 10L, method = "qml", resid.fc.FUN = NULL, ...)
}
\arguments{
\item{object}{object of class 'dfm'.}

\item{h}{forecast horizon.}

\item{method}{character. \code{"qml"}, \code{"twostep"} or \code{"pca"}.}

\item{resid.fc.FUN}{function to forecast residuals.}
}
\description{
DFM Forecasts
}
