% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.dfm}
\alias{residuals.dfm}
\title{Residuals from DFM}
\usage{
\method{residuals}{dfm}(
  object,
  method = "qml",
  orig.format = FALSE,
  standardized = FALSE,
  ...
)
}
\arguments{
\item{object}{object of class 'dfm'.}

\item{method}{character. \code{"qml"}, \code{"twostep"} or \code{"pca"}.}

\item{orig.format}{logical. \code{TRUE} returns residuals in a data format similar to \code{X}.}

\item{standardized}{logical. \code{FALSE} will put residuals on the original data scale.}
}
\description{
Residuals from DFM
}
