% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dfm}
\alias{summary.dfm}
\title{Summary information on dynamic factor model estimation}
\usage{
\method{summary}{dfm}(object, method = "qml", ...)
}
\arguments{
\item{x}{An object of \code{dfm} class}
}
\value{
Prints out a summary information following a dynamic factor
model estimation. Also can return summary plots.
}
\description{
Summary information on dynamic factor model estimation
}
