% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tsremimpNA}
\alias{tsremimpNA}
\title{Remove and Impute Missing Values in a Multivariate Time Series}
\usage{
tsremimpNA(
  X,
  max.missing = 0.5,
  na.rm.method = c("LE", "all"),
  na.impute = c("median", "rnrom", "med_MA", "med_MA_spline"),
  na.impute.MA = 3L
)
}
\arguments{
\item{X}{a matrix or multivariate time series where each column is a series.}

\item{max.missing}{proportion of series missing for a case to be considered missing.}

\item{na.rm.method}{method to apply concerning missing cases selected through \code{max.missing}: \code{"LE"} only removes cases at the beginning or end of the sample, whereas \code{"all"} always removes missing cases.}

\item{na.impute}{method to impute missing values for the PCA estimates used to initialize the EM algorithm. Note that data are standardized (scaled and centered) beforehand. Available options are:
\tabular{llll}{
\code{"median"} \tab\tab simple series-wise median imputation. \cr\cr
\code{"rnrom"} \tab\tab imputation with random numbers drawn from a standard normal distribution. \cr\cr
\code{"med_MA"} \tab\tab values are initially imputed with the median, but then a moving average is applied to smooth the estimates. \cr\cr
\code{"med_MA_spline"} \tab\tab "internal" missing values (not at the beginning or end of the sample) are imputed using a cubic spline, whereas missing values at the beginning and end are imputed with the median of the series and smoothed with a moving average.\cr\cr
}}

\item{na.impute.MA}{the order of the (2-sided) moving average applied in \code{na.impute} methods \code{"med_MA"} and \code{"med_MA_spline"}.}
}
\value{
A list with the imputed matrix \code{X_imp}, a missingness matrix \code{W} matching the dimensions of \code{X_imp},
and a vector or cases \code{na.rm} indicating cases with too many missing values that were removed beforehand.
}
\description{
This function imputes missing values in a stationary multivariate time series using various
methods, and removes cases with too many missing values.
}
