% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RcppExports.R
\name{KalmanFilter}
\alias{KalmanFilter}
\title{Implementation of a Kalman filter}
\usage{
KalmanFilter(X, A, C, Q, R, F0, P0)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{A}{Transition matrix (rp x rp)}

\item{C}{Observation matrix (n x rp)}

\item{Q}{State covariance (rp x rp)}

\item{R}{Observation covariance (n x n)}

\item{F0}{Initial state vector (rp x 1)}

\item{P0}{Initial state covariance (rp x rp)}
}
\value{
Predicted and filtered state vectors and covarainces, including a prediction for period T+1.
\tabular{lll}{
F \tab\tab T x rp filtered state vectors \cr\cr
P \tab\tab rp x rp x T filtered state covariances \cr\cr
F_pred \tab\tab T+1 x rp predicted state vectors, prediction for period t = 1 is F0 \cr\cr
P_pred \tab\tab rp x rp x T+1 predicted state covariances, prediction for period t = 1 is P0 \cr\cr
loglik \tab\tab value of the log likelihood
}
}
\description{
Implementation of a Kalman filter
}
