% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/my_RcppExports.R
\name{KalmanFilter}
\alias{KalmanFilter}
\title{Implementation of a Kalman filter}
\usage{
KalmanFilter(X, A, C, Q, R, F0, P0)

KalmanFilter(X, A, C, Q, R, F0, P0)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{A}{Transition matrix (rp x rp)}

\item{C}{Observation matrix (n x rp)}

\item{Q}{State covariance (rp x rp)}

\item{R}{Observation covariance (n x n)}

\item{F0}{Initial state vector (rp x 1)}

\item{P0}{Initial state covariance (rp x rp)}
}
\description{
Implementation of a Kalman filter

Implementation of a Kalman filter
}
