% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KalmanFilterSmoother}
\alias{KalmanFilterSmoother}
\title{Kalman Filter and Smoother}
\usage{
KalmanFilterSmoother(y, H, Q, R, F, x0, P0)
}
\arguments{
\item{y}{Data matrix (T x n)}

\item{Q}{State covariance}

\item{R}{Observation covariance}

\item{x0}{Initial state vector}

\item{P0}{Initial state covariance}

\item{C}{Observation matrix}

\item{A}{Transition matrix}
}
\description{
Kalman Filter and Smoother
}
