% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RcppExports.R
\name{KalmanFilterSmoother}
\alias{KalmanFilterSmoother}
\title{Kalman Filter and Smoother}
\usage{
KalmanFilterSmoother(X, A, C, Q, R, F0, P0)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{A}{Transition matrix (rp x rp)}

\item{C}{Observation matrix (n x rp)}

\item{Q}{State covariance (rp x rp)}

\item{R}{Observation covariance (n x n)}

\item{F0}{Initial state vector (rp x 1)}

\item{P0}{Initial state covariance (rp x rp)}
}
\value{
All results from \code{\link{KalmanFilter}} and \code{\link{KalmanSmoother}}, and additionally
a rp x rp x T matrix \code{PPm_smooth}, which is equal to the estimate of Cov(F_smooth_t, F_smooth_t-1|T) and needed for EM iterations.
}
\description{
Kalman Filter and Smoother
}
