% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RcppExports.R
\name{KalmanSmoother}
\alias{KalmanSmoother}
\title{Runs a Kalman smoother}
\usage{
KalmanSmoother(A, F, F_pred, P, P_pred)
}
\arguments{
\item{A}{Transition matrix (rp x rp)}

\item{F}{State estimates (T x rp)}

\item{F_pred}{State predicted estimates (T x rp) or (T+1 x rp)}

\item{P}{Variance estimates (rp x rp x T)}

\item{P_pred}{Predicted variance estimates (rp x rp x T) or (rp x rp x T+1)}
}
\value{
Smoothed state and covariance estimates, including initial (t = 0) values.
\tabular{lll}{
F_smooth \tab\tab T x rp smoothed state vectors, equal to the filtered state in period T \cr\cr
P_smooth \tab\tab rp x rp x T smoothed state covariance, equal to the filtered covariance in period T \cr\cr
F_smooth_0 \tab\tab 1 x rp initial smoothed state vectors, based on F0 \cr\cr
P_smooth_0 \tab\tab rp x rp initial smoothed state covariance, based on P0
}
}
\description{
Runs a Kalman smoother
}
