% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/my_RcppExports.R
\name{KalmanSmoother}
\alias{KalmanSmoother}
\title{Runs a Kalman smoother}
\usage{
KalmanSmoother(A, C, R, ZTf, ZTp, VTf_v, VTp_v)

KalmanSmoother(A, C, R, ZTf, ZTp, VTf_v, VTp_v)
}
\arguments{
\item{A}{Transition matrix (rp x rp)}

\item{C}{Observation matrix (n x rp)}

\item{R}{Observation covariance (n x n)}

\item{ZTf}{State estimates}

\item{ZTp}{State predicted estimates}

\item{VTf_v}{Variance estimates}

\item{VTp_v}{Predicted variance estimates}
}
\value{
List of smoothed estimates

List of smoothed estimates
}
\description{
Runs a Kalman smoother

Runs a Kalman smoother
}
