% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.dfm}
\alias{plot.dfm}
\title{Plot DFM}
\usage{
\method{plot}{dfm}(x, method = "qml", type = c("joint", "individual", "residual"), ...)
}
\arguments{
\item{x}{an object class 'dfm'.}

\item{method}{character. The factor estimates to use: one of \code{"qml"}, \code{"twostep"} or \code{"pca"}.}

\item{type}{character. The type of plot: \code{"joint"}, \code{"individual"} or \code{"residual"}.}
}
\description{
Plot DFM
}
