% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.dfm}
\alias{residuals.dfm}
\alias{resid.dfm}
\alias{fitted.dfm}
\title{DFM Residuals and Fitted Values}
\usage{
\method{residuals}{dfm}(
  object,
  method = switch(object$em.method, none = "twostep", "qml"),
  orig.format = FALSE,
  standardized = FALSE,
  ...
)

\method{fitted}{dfm}(
  object,
  method = switch(object$em.method, none = "twostep", "qml"),
  orig.format = FALSE,
  standardized = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class 'dfm'.}

\item{method}{character. The factor estimates to use: one of \code{"qml"}, \code{"twostep"} or \code{"pca"}.}

\item{orig.format}{logical. \code{TRUE} returns residuals/fitted values in a data format similar to \code{X}.}

\item{standardized}{logical. \code{FALSE} will put residuals/fitted values on the original data scale.}

\item{\dots}{not used.}
}
\description{
DFM Residuals and Fitted Values
}
