% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_RcppExports.R
\name{SKFS}
\alias{SKFS}
\title{(Fast) Stationary Kalman Filter and Smoother}
\usage{
SKFS(X, A, C, Q, R, F_0, P_0, loglik = FALSE)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{A}{Transition matrix (rp x rp)}

\item{C}{Observation matrix (n x rp)}

\item{Q}{State covariance (rp x rp)}

\item{R}{Observation covariance (n x n)}

\item{F_0}{Initial state vector (rp x 1)}

\item{P_0}{Initial state covariance (rp x rp)}

\item{loglik}{logical. Compute log-likelihood?}
}
\value{
All results from \code{\link{SKF}} and \code{\link{FIS}}, and additionally
a rp x rp x T matrix \code{PPm_smooth}, which is equal to the estimate of \eqn{Cov(F^smooth_t, F^smooth_{t-1} | T)}{Cov(F_smooth(t), F_smooth(t-1) | T)} and needed for EM iterations.
}
\description{
(Fast) Stationary Kalman Filter and Smoother
}
\seealso{
\code{\link{SKF}} \code{\link{FIS}}
}
