% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dfm}
\alias{summary.dfm}
\alias{print.dfm}
\alias{print.dfm_summary}
\title{DFM Summary Methods}
\usage{
\method{print}{dfm}(x, digits = 4L, ...)

\method{summary}{dfm}(
  object,
  method = switch(object$em.method, none = "twostep", "qml"),
  ...
)

\method{print}{dfm_summary}(x, digits = 4L, compact = sum(x$info["n"] > 15, x$info["n"] > 40), ...)
}
\arguments{
\item{x, object}{an object class 'dfm'.}

\item{digits}{integer. The number of digits to print out.}

\item{\dots}{not used.}

\item{method}{character. The factor estimates to use: one of \code{"qml"}, \code{"twostep"} or \code{"pca"}.}

\item{compact}{integer. Display a more compact printout: \code{0} prints everything, \code{1} omits the observation matrix [C] and covariance matrix [R], and \code{2} omits all disaggregated information - yielding a summary of only the factor estimates.}
}
\value{
Summary information following a dynamic factor model estimation.
}
\description{
Summary and print methods for class 'dfm'. \code{print.dfm} just prints basic model information and the factor transition matrix [A],
\code{summary.dfm} returns all system matrices and additional residual and goodness of fit statistics - with a print method allowing full or compact printout.
}
